requirements.txt
tensorflow
streamlit
Pillow
numpy
matplotlib